updated Float of numbers to Int

+++ Cells/DiaryEntryTableViewCell.swift
@@ -24,13 +24,13 @@ class DiaryEntryTableViewCell: UITableViewCell {
         format.dateFormat = "MM/dd/yyyy 'at' HH:mm"
 
         //Every time item is updated our tableview Will be too, but this usually won't be necessary in the current incantation of the app
-        entry.rx.observe(Float.self, "o3")
+        entry.rx.observe(Int.self, "o3")
             .subscribe(onNext: { [weak self] o3 in
                 self?.o3AQI.text = String(o3!)
             })
             .disposed(by: disposeBag)
         
-        entry.rx.observe(Float.self, "pm25")
+        entry.rx.observe(Int.self, "pm25")
             .subscribe(onNext: { [weak self] pm25 in
                 self?.PM25AQI.text = String(pm25!)
             })
